more info on decoded range, raw data by offset
mptcore
  save offset and decoded data length in decoder state
  extra entry for message length (at beginning of decoded range)
  only save current raw data offset in decoder state
  change return type of decoder functions (all sizes in state)
  prune unused space when decoding queue content

mptio
  no data decoding in stream poll operation
  replace manual prune operations with decoder queue shift call

+++ mptcore/convert.h
@@ -192,20 +192,20 @@ extern const char *mpt_encoding_type(int);
 /* standard COBS(/R) operations */
 extern ssize_t mpt_encode_cobs(MPT_STRUCT(encode_state) *, const struct iovec *, const struct iovec *);
 extern ssize_t mpt_encode_cobs_r(MPT_STRUCT(encode_state) *, const struct iovec *, const struct iovec *);
-extern ssize_t mpt_decode_cobs(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
-extern ssize_t mpt_decode_cobs_r(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
+extern int mpt_decode_cobs(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
+extern int mpt_decode_cobs_r(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
 extern int mpt_cobs_dec(const void *, size_t *, void *, size_t *);
 /* COBS/ZPE(+R) operations */
 extern ssize_t mpt_encode_cobs_zpe(MPT_STRUCT(encode_state) *, const struct iovec *, const struct iovec *);
 extern ssize_t mpt_encode_cobs_zpe_r(MPT_STRUCT(encode_state) *, const struct iovec *, const struct iovec *);
-extern ssize_t mpt_decode_cobs_zpe(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
-extern ssize_t mpt_decode_cobs_zpe_r(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
+extern int mpt_decode_cobs_zpe(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
+extern int mpt_decode_cobs_zpe_r(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
 extern int mpt_cobs_dec_zpe(const void *, size_t *, void *, size_t *);
 
 /* filter non-zero inline data */
 extern ssize_t mpt_encode_string(MPT_STRUCT(encode_state) *, const struct iovec *, const struct iovec *);
 /* add command header to zero-terminated text */
-extern ssize_t mpt_decode_command(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
+extern int mpt_decode_command(MPT_STRUCT(decode_state) *, const struct iovec *, size_t);
 
 
 /* convert structured data to string */
